+++ src/draw/draw.hpp
@@ -7,8 +7,8 @@
 #include <string>
 #include <vector>
 
-#include "../math/float2.hpp"
-#include "../math/float3.hpp"
+#include "../math/double2.hpp"
+#include "../math/double3.hpp"
 #include "../math/vector.hpp"
 #include "../math/triangle.hpp"
 
@@ -17,7 +17,7 @@
 namespace draw {
 namespace BMP {
 
-void Write(const std::vector<std::vector<math::float3>>& image, const std::string& name);
+void Write(const std::vector<std::vector<math::double3>>& image, const std::string& name);
 
 }
 }
